+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.6.3
 * Fix Windows toolbar icon
+* Show progress when downloading DCS-BIOS or new DCSpy release
 
 ## 3.6.2
 * Initial support colors for G19 - @emcek